Updated some of the high availability migrations in order for some tables to contain date columns. Added new tables (migrations) and models for the high availability snapshot, members and members after consensus.

+++ migrations/2021-12-13-183554_create_high_avail_storage/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 create table ha_state_machine (
   id nvarchar(50) primary key not null,
-  last_applied_log NUMERIC not null,
+  last_applied_log int not null,
   date_created nvarchar(100) not null,
   date_updated nvarchar(100) not null
 );
